Fix game reset style
Instead of reloading the page after first to five has been declared,
this modification just returns the dom back to the initial state it was
when the it was first requested by client by storing the innerHTML in
a variable which is then used by a function and the function is in turn
called by the 'click' event set on the reset button. Selah!